requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==3.1.0
Django==4.1.7
djangorestframework==3.14.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pytz==2022.7.1
slackclient==2.9.4
sqlparse==0.4.3
yarl==1.8.2
